+++ packages/AI/Vectors/Sync/src/generic/vectorSync.types.ts
@@ -11,7 +11,7 @@ export type VectorizeEntityParams = {
      * If defined, all records within the given list will be vectorized
      * instead of all records within the entity. Note that the list's
      * EntityID field must equal the entityID parameter.
-     */
+    */
     listID?: string;
     /**
      * The number of records to be vectorized and inserted
@@ -75,21 +75,3 @@ export type AnnotateWorkerContext = {
     ParamName: string;
     Data: unknown[];
   };
-
-  export type CreateEntityDocumentParams = {
-    EntityID: string,
-    VectorDatabaseID: string,
-    TemplateID?: string,
-    AIModelID: string,
-    VectorDatabaseDetails?: {
-      Name: string,
-      Description: string | null,
-      DefaultURL: string,
-      ClassKey: string,
-    },
-    TemplateDetails?: {
-      Name: string,
-      Description: string | null,
-      TemplateContent: string,
-    }
-  }