+++ packages/MJCore/src/generic/baseEntity.ts
@@ -483,10 +483,10 @@ export abstract class BaseEntity<T = unknown> {
      * @returns 
      */
     public GetFieldByName(fieldName: string): EntityField | null {
-        if(!fieldName){
+        if(!fieldName) {
             return null;
         }
-
+        
         const lcase = fieldName.trim().toLowerCase(); // do this once as we will use it multiple times
         return this.Fields.find(f => f.Name.trim().toLowerCase() === lcase);
     }