+++ packages/AI/Vectors/Database/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Change Log - @memberjunction/ai-vectordb
 
-This log was last generated on Sun, 08 Sep 2024 19:33:23 GMT and should not be manually modified.
+This log was last generated on Fri, 20 Sep 2024 17:51:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.5.1
+
+Fri, 20 Sep 2024 17:51:58 GMT
+
+### Patches
+
+- Bump @memberjunction/core to v2.5.1
+- Bump @memberjunction/global to v2.5.1
+
+## 2.5.0
+
+Fri, 20 Sep 2024 16:17:07 GMT
+
+### Minor changes
+
+- Bump minor version (craig.adam@bluecypress.io)
+- Bump @memberjunction/core to v2.5.0
+- Bump @memberjunction/global to v2.5.0
 
 ## 2.4.1