+++ packages/Angular/Generic/timeline/CHANGELOG.md
@@ -1,8 +1,35 @@
 # Change Log - @memberjunction/ng-timeline
 
-This log was last generated on Sun, 08 Sep 2024 19:33:23 GMT and should not be manually modified.
+This log was last generated on Fri, 20 Sep 2024 17:51:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.5.1
+
+Fri, 20 Sep 2024 17:51:58 GMT
+
+### Patches
+
+- Bump @memberjunction/core-entities to v2.5.1
+- Bump @memberjunction/global to v2.5.1
+- Bump @memberjunction/core to v2.5.1
+- Bump @memberjunction/ng-container-directives to v2.5.1
+- Bump @memberjunction/ng-entity-form-dialog to v2.5.1
+- Bump @memberjunction/ng-shared to v2.5.1
+
+## 2.5.0
+
+Fri, 20 Sep 2024 16:17:06 GMT
+
+### Minor changes
+
+- Bump minor version (craig.adam@bluecypress.io)
+- Bump @memberjunction/core-entities to v2.5.0
+- Bump @memberjunction/global to v2.5.0
+- Bump @memberjunction/core to v2.5.0
+- Bump @memberjunction/ng-container-directives to v2.5.0
+- Bump @memberjunction/ng-entity-form-dialog to v2.5.0
+- Bump @memberjunction/ng-shared to v2.5.0
 
 ## 2.4.1